226117ing

+++ code/codeforces/833B/833B.cpp
@@ -117,7 +117,7 @@ void _solve(int l,int r,int mn,int mx) {
     }
     int m=l+r>>1;
     int id=0,lim=min(mx,m-1);
-    for(int16 i=mn;i<=lim;i++) {
+    for(int i=mn;i<=lim;i++) {
         int16 nv=dp[i]+seg.qu(i+1,m);
         if(nv>ndp[m]) {
             ndp[m]=nv;
